+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/wrappers/EyesRemoteWebElement.java
@@ -12,7 +12,10 @@ import org.openqa.selenium.*;
 import org.openqa.selenium.interactions.Coordinates;
 import org.openqa.selenium.remote.*;
 
+import java.lang.reflect.Field;
+import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
+import java.lang.reflect.Proxy;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -108,6 +111,8 @@ public class EyesRemoteWebElement extends RemoteWebElement {
         this.logger = logger;
         this.eyesDriver = eyesDriver;
 
+
+        webElement = getWrappedWebElement(webElement);
         if (webElement instanceof RemoteWebElement) {
             this.webElement = (RemoteWebElement) webElement;
         } else {
@@ -127,6 +132,32 @@ public class EyesRemoteWebElement extends RemoteWebElement {
         } catch (NoSuchMethodException e) {
             throw new EyesException("Failed to find 'execute' method!");
         }
+    }
+
+    /**
+     * If the web element was created by {@link org.openqa.selenium.support.FindBy}, then it's a {@link Proxy} object.
+     * This method gets the real web element from the proxy object.
+     */
+    static WebElement getWrappedWebElement(WebElement webElement) {
+        if (!(webElement instanceof Proxy)) {
+            return webElement;
+        }
+
+        Proxy proxy = (Proxy) webElement;
+        Field[] fields =  Proxy.class.getDeclaredFields();
+        for (Field field : fields) {
+            if(field.getType().equals(InvocationHandler.class)) {
+                field.setAccessible(true);
+                try {
+                    InvocationHandler handler = (InvocationHandler) field.get(proxy);
+                    return  (WebElement) handler.invoke(null, WrapsElement.class.getMethod("getWrappedElement"), null);
+                } catch (Throwable throwable) {
+                    throw new EyesException("Failed getting web element from page object", throwable);
+                }
+            }
+        }
+
+        throw new IllegalStateException("InvocationHandler field wasn't found in proxy class");
     }
 
     public Region getBounds() {