+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Tests now won't run forever after a failed rendering in the VG server. [Trello 1907](https://trello.com/c/n80nncwf)
 ### Updated
 - Updated browser info api as required. [Trello 1872](https://trello.com/c/bykk2rzB)
+### Added
+- Supporting tests page factory object. [Trello 1503](https://trello.com/c/pjmn2N7H)
 
 ## [3.165.0]
 ### Added